fix: address review feedback on package reorganization
- Add missing artifact cache clear in ingest_main() between datasets
- Update stale imports in tests/conftest.py and tests/test_data_retrieval.py
- Fix extra blank line in utils.py
